Frontend & Backend Initialized
Changes Made:
Frontend: Next.JS framework installed as the name app-dashboard. Component: sidebar.tsx started on.

Backend : Bot was initialized with the format that will be used created. Routes the bring the API over from the .env (not available on github) was created. A basic strategy to know when to buy and sell was worked on, MODIFY LATER.